+++ internal/cmd/rootcmd/rootcmd.go
@@ -2,9 +2,13 @@
 package rootcmd
 
 import (
+	"encoding/json"
+	"fmt"
+
 	"github.com/spf13/cobra"
 
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/manager"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/manager/metadata"
 )
 
 // Execute is the entry point to the controller manager.
@@ -17,8 +21,33 @@ func Execute() {
 				return Run(cmd.Context(), &cfg)
 			},
 			SilenceUsage: true,
+			// We can silence the errors because cobra.CheckErr below will print
+			// the returned error and set the exit code to 1.
+			SilenceErrors: true,
+		}
+		versionCmd = &cobra.Command{
+			Use:   "version",
+			Short: "Show JSON version information",
+			RunE: func(cmd *cobra.Command, args []string) error {
+				type Version struct {
+					Release string `json:"release"`
+					Repo    string `json:"repo"`
+					Commit  string `json:"commit"`
+				}
+				out, err := json.Marshal(Version{
+					Release: metadata.Release,
+					Repo:    metadata.Repo,
+					Commit:  metadata.Commit,
+				})
+				if err != nil {
+					return fmt.Errorf("failed to print version information: %w", err)
+				}
+				fmt.Printf("%s\n", out)
+				return nil
+			},
 		}
 	)
+	rootCmd.AddCommand(versionCmd)
 	rootCmd.Flags().AddFlagSet(cfg.FlagSet())
 	cobra.CheckErr(rootCmd.Execute())
 }