+++ test/e2e/features_test.go
@@ -341,6 +341,7 @@ func TestDeployAllInOneDBLESSGateway(t *testing.T) {
 	t.Log("running the admission webhook setup script")
 	deployAdmissionWebhook(t, env)
 
+
 	// vov it's easier than tracking the deployment state
 	t.Log("creating a consumer to ensure the admission webhook is online")
 	consumer := &kongv1.KongConsumer{