+++ test/e2e/utils_test.go
@@ -186,7 +186,10 @@ func getTestManifest(t *testing.T, baseManifestPath string) (io.Reader, error) {
 func getCurrentGitTag(path string) (semver.Version, error) {
 	cmd := exec.Command("git", "describe", "--tags")
 	cmd.Dir = path
-	tagBytes, _ := cmd.Output()
+	tagBytes, err := cmd.Output()
+	if err != nil {
+		return semver.Version{}, fmt.Errorf("%q command failed: %w", cmd.String(), err)
+	}
 	tag, err := semver.ParseTolerant(string(tagBytes))
 	if err != nil {
 		return semver.Version{}, err
@@ -315,17 +318,17 @@ func startPortForwarder(
 	namespace, name string,
 	localPort, targetPort string,
 ) {
+
 	kubeconfig, cleanup := getTemporaryKubeconfig(t, env)
 	defer cleanup()
 
 	cmd := exec.CommandContext(ctx, "kubectl", "--kubeconfig", kubeconfig, "port-forward", "-n", namespace,
 		name, fmt.Sprintf("%s:%s", localPort, targetPort),
 	)
+
 	t.Logf("forwarding port %s to %s/%s:%s", localPort, namespace, name, targetPort)
 	if startErr := cmd.Start(); startErr != nil {
-		startOutput, outputErr := cmd.Output()
-		assert.NoError(t, outputErr)
-		require.NoError(t, startErr, string(startOutput))
+		require.NoError(t, startErr, out.String())
 	}
 	require.Eventually(t, func() bool {
 		conn, err := net.Dial("tcp", fmt.Sprintf("localhost:%s", localPort))
@@ -333,6 +336,8 @@ func startPortForwarder(
 			conn.Close()
 			return true
 		}
+
+		t.Logf("port forwarding command %q output so far: %s", cmd.String(), out.String())
 		return false
 	}, kongComponentWait, time.Second)
 }